feat(mongo3.6): customize mongo logger
This patch customizes the mongo logger configuration to replace the
default configuration in fluentd-catch-all-config. The
mongo.logger.conf file describes the parser of the log of
mongo(/var/log/mongodb/mongod.log) and mutates the severity string to
corresponding log entry severity. Replace the log configuration of
default fluentd with the customized one if the service is
actived.(i.e, stackdriver-agent is installed).